+++ frontend/src/GymCard.js
@@ -1,5 +1,6 @@
 import styled from "styled-components"
 import { MdDone, MdTurnedInNot } from "react-icons/md"
+import { useState } from 'react'
 
 export default function GymCard ({gyms}) {
 
@@ -64,42 +65,88 @@ export default function GymCard ({gyms}) {
 
     const ButtonStyled = styled.button`
         
-        background-color: #FD5B3F;
-        width: auto;
-        height: 40px;
-        box-shadow: 0 0 30px 0px rgb(66, 75, 84, 0.2);
-        border-radius: 10px;
-        border-style: none;
-        color: #E9E9E9;
-        position: absolute;
-        bottom: 8px;
-        right: 8px;
+            background-color: #FD5B3F;
+            width: auto;
+            height: 40px;
+            box-shadow: 0 0 30px 0px rgb(66, 75, 84, 0.2);
+            border-radius: 10px;
+            border-style: none;
+            color: #E9E9E9;
+            position: absolute;
+            bottom: 8px;
+            right: 8px;
         `
 
     const GymCard = styled.div`
 
-        position: relative;
-        background-color: white;
-        margin: auto;
-        width: 90%;
-        box-shadow: 0 0 30px 0px rgb(66, 75, 84, 0.2);
-        border-radius: 10px;
-        font-family: "Roboto";
+            position: relative;
+            background-color: white;
+            margin: auto;
+            width: 90%;
+            box-shadow: 0 0 30px 0px rgb(66, 75, 84, 0.2);
+            border-radius: 10px;
+            font-family: "Roboto";
 
         `
 
     const ButtonBookmark = styled.div`
 
-        position: absolute;
-        right: 8px;
-        top: 8px;
-        font-size: 35px;
-        color: #FD5B3F;
+            position: absolute;
+            right: 8px;
+            top: 8px;
+            font-size: 35px;
+            color: #FD5B3F;
 
         `
+    
+    const SortFilterStyled = styled.div`
+
+            background-color: white;
+            margin: auto;
+            width: 90%;
+            box-shadow: 0 0 30px 0px rgb(66, 75, 84, 0.2);
+            border-radius: 10px;
+            font-family: "Roboto";
+            display: flex;
+            justify-content: space-evenly;
+        `
+
+    const SortStyled = styled.div`
+            flex: 1;
+            border-right: 1px solid black;
+            display: flex;
+            justify-content: center;
+            margin: 5px 0px;
+            `
+
+    const FilterStyled = styled.div`
+            flex: 1;
+            display: flex;
+            justify-content: center;
+            margin: 5px 0px;
+            `
+
+    const [sort, setSort] = useState(false)
+    
+
+    let SortedGymsAlphabetAsc = gyms.sort((a,b) => (a.name > b.name) ? 1 : -1)
+    let SortedGymsAlphabetDesc = gyms.sort((a,b) => (a.name < b.name) ? 1 : -1)
+
+    let SortedGymsMonthlyPriceAsc = gyms.sort((a,b) => (a.monthlyPrice > b.monthlyPrice) ? 1 : -1)
+    let SortedGymsMonthlyPriceDesc = gyms.sort((a,b) => (a.monthlyPrice > b.monthlyPrice) ? 1 : -1)
+
+    let SaunaGyms = gyms.filter(gym => gym.sauna === true)
 
     return (
-        gyms.map(({name, registrationFee, monthlyPrice, wellness, parking, childCare, personalTraining}) =>(
+        <>
+        <SortFilterStyled>
+            <SortStyled onCLick={() => setSort(true)}>Sortieren</SortStyled>
+            <FilterStyled>Filtern</FilterStyled>
+        </SortFilterStyled>
+        {sort && (
+            <div>Hallo</div>
+        )}
+        {gyms.map(({name, registrationFee, monthlyPrice, wellness, parking, childCare, personalTraining}) =>(
             <GymCard>
             <Name>{name}</Name>
             <ButtonBookmark><MdTurnedInNot /></ButtonBookmark>
@@ -111,8 +158,10 @@ export default function GymCard ({gyms}) {
             <MonthlyPrice>Monatspreis: <span style={{color: "red"}}>{monthlyPrice / Math.pow(10, 2)}€</span></MonthlyPrice>
             <ButtonStyled>Zum Anbieter</ButtonStyled>
             </GymCard>
-            )) 
+            )) }
+            </>
     )
+
 
 
 }
